refactor(core): extract props parsing to separate file
